requirements.txt
flask
flask-restplus
requests
pyyaml
